feat: 게시글 관리 페이지 수정
- 게시글 관리 페이지 수정

#43

+++ src/views/PostAdmin.vue
@@ -4,7 +4,7 @@
         <button class="btn opacity-0 cursor-default">zzz</button>
         <button v-if="checkList != ''" class="btn btn-active btn-ghost" @click="deleteListFunc">삭제</button>
     </div>
-    <div class="px-16 mt-6 flex justify-end">
+    <div class="px-16 mt-6 flex justify-start">
         <div class="flex">
             <div class="input-group w-fit border-2 rounded-xl mr-4">
                 <select class="select" v-model="requestData.keywordType">
@@ -32,15 +32,15 @@
                            v-model="checkBoolean"/>
                 </th>
                 <th class="bg-base-100">
-                    <button v-if="requestData.column !== 'postId'" @click="handleSort('postId')" class="flex">등록번호
+                    <button v-if="requestData.column !== 'member.username'" @click="handleSort('member.username')" class="flex">이름
                         <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" class="w-5 h-5 opacity-0">
                             <path fill-rule="evenodd"
                                   d="M10 15a.75.75 0 01-.75-.75V7.612L7.29 9.77a.75.75 0 01-1.08-1.04l3.25-3.5a.75.75 0 011.08 0l3.25 3.5a.75.75 0 11-1.08 1.04l-1.96-2.158v6.638A.75.75 0 0110 15z"
                                   clip-rule="evenodd"/>
                         </svg>
                     </button>
-                    <button v-if="requestData.column === 'postId'" class="text-primary flex" @click="handleSort('postId')">
-                        등록번호
+                    <button v-if="requestData.column === 'member.username'" class="text-primary flex" @click="handleSort('member.username')">
+                        이름
                         <svg v-if="requestData.sortDirection === 'asc'" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"
                              fill="currentColor" class="w-5 h-5">
                             <path fill-rule="evenodd"
@@ -54,9 +54,6 @@
                                   clip-rule="evenodd"/>
                         </svg>
                     </button>
-                </th>
-                <th class="bg-base-100">
-                    이름
                 </th>
                 <th class="bg-base-100">
                     <button v-if="requestData.column !== 'title'" @click="handleSort('title')" class="flex">제목
@@ -106,6 +103,30 @@
                         </svg>
                     </button>
                 </th>
+                <th class="bg-base-100">
+                    <button v-if="requestData.column !== 'createdDate'" @click="handleSort('createdDate')" class="flex">작성일
+                        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" class="w-5 h-5 opacity-0">
+                            <path fill-rule="evenodd"
+                                  d="M10 15a.75.75 0 01-.75-.75V7.612L7.29 9.77a.75.75 0 01-1.08-1.04l3.25-3.5a.75.75 0 011.08 0l3.25 3.5a.75.75 0 11-1.08 1.04l-1.96-2.158v6.638A.75.75 0 0110 15z"
+                                  clip-rule="evenodd"/>
+                        </svg>
+                    </button>
+                    <button v-if="requestData.column === 'createdDate'" class="text-primary flex" @click="handleSort('createdDate')">
+                        작성일
+                        <svg v-if="requestData.sortDirection === 'asc'" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"
+                             fill="currentColor" class="w-5 h-5">
+                            <path fill-rule="evenodd"
+                                  d="M10 15a.75.75 0 01-.75-.75V7.612L7.29 9.77a.75.75 0 01-1.08-1.04l3.25-3.5a.75.75 0 011.08 0l3.25 3.5a.75.75 0 11-1.08 1.04l-1.96-2.158v6.638A.75.75 0 0110 15z"
+                                  clip-rule="evenodd"/>
+                        </svg>
+                        <svg v-if="requestData.sortDirection === 'desc'" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20"
+                             fill="currentColor" class="w-5 h-5">
+                            <path fill-rule="evenodd"
+                                  d="M10 5a.75.75 0 01.75.75v6.638l1.96-2.158a.75.75 0 111.08 1.04l-3.25 3.5a.75.75 0 01-1.08 0l-3.25-3.5a.75.75 0 111.08-1.04l1.96 2.158V5.75A.75.75 0 0110 5z"
+                                  clip-rule="evenodd"/>
+                        </svg>
+                    </button>
+                </th>
                 <th class="bg-base-100"></th>
             </tr>
             </thead>
@@ -115,9 +136,6 @@
                     <input type="checkbox" class="checkbox" v-model="checkList" :value="i.postId"/>
                 </th>
                 <td>
-                   {{i.postId}}
-                </td>
-                <td>
                     <div class="flex items-center space-x-3">
                         <div class="avatar">
                             <div class="mask mask-squircle w-12 h-12">
@@ -126,8 +144,6 @@
                         </div>
                         <div>
                             <div class="font-bold">{{ i.username }}</div>
-                            <div class="text-sm opacity-50">게시일 {{ i.createdDate }}</div>
-                            <div class="text-sm opacity-50" v-if="i.modifiedDate">수정일 {{ i.createdDate }}</div>
                         </div>
                     </div>
                 </td>
@@ -166,7 +182,7 @@
                                 <div class="modal-action">
                                     <label :for="i.postId" class="btn btn-active btn-ghost"
                                            @click="deleteOneFunc(i.postId)">삭제</label>
-                                    <label :for="i.postId" class="btn">Close!</label>
+                                    <label :for="i.postId" class="btn"></label>
                                 </div>
                             </div>
                         </label>
@@ -177,6 +193,7 @@
                     </div>
                 </td>
                 <td>{{ i.views }}</td>
+                <td>{{ i.preCreatedDate }}</td>
                 <th>
                     <button id="deleteOneStudentBtn" class="btn btn-circle btn-ghost"
                             @click="deleteOneFunc(i.postId)">
@@ -202,10 +219,11 @@ import Pagination from "../components/common/Pagination.vue";
 
 let response
 const posts = ref([]);
+const allPageCheckList = ([]);
 const checkList = ref([]);
 const checkBoolean = ref();
 let requestData = ref({
-    column: "postId",
+    column: "createdDate",
     sortDirection: "desc",
     keywordType: "title",
     keyword: "",
@@ -215,25 +233,29 @@ let requestData = ref({
 const paging = ref({});
 const load = ref(false)
 
-const init = async () => {
+const getPostList = async () => {
     response = await getList(requestData.value)
     posts.value = response.data.postList
     paging.value = response.data.pagingUtil
-    console.log(response.data.pagingUtil)
+}
+
+const init = async () => {
+    await getPostList()
     load.value = true
 }
 init()
 
 const deleteListFunc = async () => {
-    const postIds = checkList.value
-    const id = {
-        postIds
+    let answer = confirm(`${checkList.value.length}명의 학생을 삭제하시겠습니까?`)
+    if(answer){
+        const postIds = checkList.value
+        const id = {
+            postIds
+        }
+        await deleteList(id)
+        await getPostList()
+        alert("삭제가 완료되었습니다~")
     }
-    await deleteList(id)
-    console.log(requestData.value)
-    response = await getList(requestData.value)
-    posts.value = response.data.postList
-    alert("삭제가 완료되었습니다~")
 }
 
 const deleteOneFunc = async (postId) => {
@@ -244,8 +266,7 @@ const deleteOneFunc = async (postId) => {
         postIds
     }
     await deleteList(id)
-    response = await getList(requestData.value)
-    posts.value = response.data.postList
+    await getPostList()
     alert("삭제가 완료되었습니다~")
 }
 
@@ -256,19 +277,21 @@ const handleAllCheckBox = () => {
             checkList.value.push(i.postId)
         }
     }
+    console.log(checkList.value)
 }
 
 const handleDirection = async () => {
-    response = await getList(requestData.value)
-    posts.value = response.data.postList
-    paging.value = response.data.pagingUtil
+    requestData.value.page = 1
+    await getPostList()
+    if(posts.value.length == 0){
+        load.value = false
+    }
 }
 
 const changePage = async (page) => {
     requestData.value.page = page
-    response = await getList(requestData.value)
-    posts.value = response.data.postList
-    paging.value = response.data.pagingUtil
+    checkBoolean.value = false
+    await getPostList()
 }
 
 const handleSort = async (column) => {
@@ -282,9 +305,7 @@ const handleSort = async (column) => {
         requestData.value.column = column;
         requestData.value.sortDirection = 'desc';
     }
-    response = await getList(requestData.value)
-    posts.value = response.data.postList
-    paging.value = response.data.pagingUtil
+    await getPostList()
 }
 </script>
